MarkDups performance improved
